motor refactor

+++ Core/Inc/comms.h
@@ -55,9 +55,9 @@ void int16_to_bytes(int16_t value, uint8_t *buffer);
 uint16_t crc16(const uint8_t* data_p, uint8_t length);
 void UARTSendPos(TxCommsData* txCommsData);
 
-void handleCommand(RxCommsData* rxCommsData, uint8_t* setVelocity);
+void handleCommand(RxCommsData* rxCommsData, int16_t* setVelocity);
 
-void handleRx(RxCommsData* rxCommsData, uint8_t* setVelocity);
+void handleRx(RxCommsData* rxCommsData, int16_t* setVelocity);
 void handleTx(TxCommsData* txCommsData, PollTimers* pollTimers); // make pollTimers internal? call it on interrupts?
